+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - feature: install dropbox
 - feature: install obsidian
 - feature: install aider
+- feature: install bruno
 
 ## [1.4.0] - 2026-07-24